+++ src/Plugin/ParallelExecution/Event/ExecuteStart.php
@@ -13,8 +13,6 @@ use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * ExecuteStart event, see Events for details.
- *
- * @codeCoverageIgnore
  */
 class ExecuteStart extends Event
 {